Refactor: move server api to its own file
memcached.cc is already way too big

Reviewed-on: http://review.couchbase.org/119151
Tested-by: Build Bot <build@couchbase.com>

+++ daemon/settings.h
@@ -811,7 +811,7 @@ public:
         notify_changed("num_reader_threads");
     }
 
-    size_t getNumWriterThreads() const {
+    int getNumWriterThreads() const {
         return num_writer_threads.load(std::memory_order_acquire);
     }
